docs: 질문상자 부분 QuestionBox.js로 컴포넌트 분리 및 css파일 추가
