+++ src/components/QuestionBox.module.css
@@ -1,0 +1,133 @@
+/* 질문 영역 */
+.container {
+  background-color: var(--gray-scale-20);
+  padding: 189px 0 140px;
+  @media (max-width: 767px) {
+    padding: 176px 24px 140px;
+  }
+}
+
+.questions_box {
+  background-color: var(--brown-scale-10);
+  width: 100%;
+  max-width: 716px;
+  height: 100%;
+  min-height: 330px;
+  padding: 16px;
+  margin: 0 auto;
+  border: 1px solid var(--brown-scale-20);
+  border-radius: 16px;
+}
+
+.questions_box__title {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+}
+
+.questions_box__title p {
+  color: var(--brown-scale-40);
+  font-size: 20px;
+  font-weight: 400;
+  line-height: 25px;
+  @media (max-width: 767px) {
+    font-size: 18px;
+    line-height: 24px;
+  }
+}
+
+/* 질문섹션 */
+.section {
+  background-color: var(--gray-scale-10);
+  width: 100%;
+  min-height: 243px;
+  height: 100%;
+  padding: 32px;
+  margin-top: 18px;
+  border-radius: 16px;
+  box-shadow: var(--shadow-1pt);
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+}
+.section_badge,
+.active {
+  font-family: "Pretendard Variable", Pretendard, -apple-system;
+  font-size: 14px;
+  font-weight: var(--font-weight-medium);
+  line-height: 18px;
+  padding: 6px 12px 4px;
+  border-radius: 8px;
+}
+.section_badge {
+  width: 63px;
+  color: var(--gray-scale-40);
+  background-color: var(--gray-scale-10);
+  border: 1px solid var(--gray-scale-40);
+}
+.active {
+  width: 79px;
+  color: var(--brown-scale-40);
+  background-color: var(--gray-scale-10);
+  border: 1px solid var(--brown-scale-40);
+}
+
+.section_title {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.section_title__date {
+  color: var(--gray-scale-40);
+  font-family: "Pretendard Variable", Pretendard, -apple-system;
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 18px;
+}
+.section_title__content {
+  font-size: 18px;
+  font-weight: 400px;
+  line-height: 24px;
+  @media (max-width: 767) {
+    font-size: 16px;
+    line-height: 22px;
+  }
+}
+
+.section_reactions {
+  padding-top: 24px;
+  border-top: 1px solid var(--gray-scale-30);
+  display: flex;
+  gap: 32px;
+}
+.section_reactions div {
+  color: var(--gray-scale-40);
+  font-family: "Pretendard Variable", Pretendard, -apple-system;
+  font-size: 14px;
+  font-weight: var(--font-weight-medium);
+  line-height: 18px;
+  display: flex;
+  justify-content: center;
+  gap: 6px;
+}
+
+/* 질문없을 때, 빈페이지 */
+.empty_img {
+  width: 150px;
+  height: 154px;
+  display: block;
+  margin: 68px auto;
+  @media (max-width: 767px) {
+    width: 114px;
+    height: 118px;
+  }
+}
+
+@media (max-width: 767px) {
+  .questions_box__title img {
+    width: 22px;
+    height: 22px;
+  }
+}
